requirements.txt
django==4.1.6
djangorestframework==3.14.0
drf-yasg==1.21.5
django-debug-toolbar==3.8.1